refactor(test): P5-4 extract pulp pr shell-out tests
Phase 5 P5-4 follow-up. `pulp pr` shell-out tests extracted from
test_cli_shellout.cpp into test_cli_shellout_pr.cpp (12 TEST_CASEs,
12 assertions, 350 lines).

Two contiguous clusters bundled into one extracted TU:
  * pr.workflow validation (manual / github / shipyard delegation)
  * Shipyard binary pin handling
  * Native fallback path (no shipyard installed)
  * pulp #643 install guidance, native mode validation, native help

test_cli_shellout.cpp: 2,646 → 2,314 (-332 lines, closer to < 2,000
P5-4 target).
